+++ Tests/VCoreTests/Tests/Helpers/VCoreErrorTests.swift
@@ -1,0 +1,22 @@
+//
+//  VCoreErrorTests.swift
+//  VCoreTests
+//
+//  Created by Vakhtang Kontridze on 10.05.22.
+//
+
+import XCTest
+@testable import VCore
+
+// MARK: - Tests
+final class VCoreErrorTests: XCTestCase {
+    func test() {
+        XCTAssertEqual(SomeError.errorDomain, "com.vcore")
+    }
+    
+    private struct SomeError: VCoreError {
+        var domain: String
+        var code: Int
+        var description: String
+    }
+}